Started the profile page
Added a profile page, and a button component.
So far profile page display the user's name, username and two buttons to
add a note and add a class.
Below this is the note area, which I think would be cool to eventually
display the user's notes, maybe a drag and drop upload option.
TODO: find something to fill out the space a bit more. Add an edit
button.